+++ CHANGELOG.md
@@ -12,6 +12,10 @@ vim.keymap.set('i', '<C-z>', '<C-g>u<Esc>[s1z=`]a<C-g>u', { desc = 'Correct late
 - FEATURE: Add `tab:> ` to 'listchars' option when `options.extra_ui` is set. This prevents showing `^I` instead of a tab and actual value comes from Neovim's default.
 - FEATURE: Set `termguicolors` only on Neovim<0.10, as later versions should have it on by default (if terminal emulator supports it).
 
+## mini.comment
+
+- FEATURE: Hooks are now called with data about commenting action.
+
 ## mini.deps
 
 Introduction of a new module.